examples/simple_python_websocket/requirements.txt
requests==2.25.1
python-dotenv==0.16.0